refactor(api): promote election_group_id from SMALLINT to INT across all schema layers
This change upgrades election_groups.id from SMALLINT to INT (int32) and propagates
the wider type through all foreign key columns, sqlc-generated Go models, query
params, services, handlers, and background workers. The previous SMALLINT was an
over-optimisation that caused cascading int16/int32 type mismatches throughout the
codebase and limited future scalability.

## Database migrations
- election_groups.id: SMALLINT -> INT
- elections.election_group_id: SMALLINT -> INT
- party_election_groups.election_group_id: SMALLINT -> INT
- polling_unit_assignments.election_group_id: SMALLINT -> INT
- supervisor_assignments (state/lga/ward): election_group_id SMALLINT -> INT
- party_applications.election_group_id: SMALLINT -> INT
- polling_unit_updates.election_group_id: SMALLINT -> INT
- polling_unit_results / unmatched_polling_unit_results: SMALLINT -> INT
- election_votes / did_not_vote_reasons: SMALLINT -> INT
- election_stats tables (all geo-level tables): SMALLINT -> INT
- party_marketing_campaigns.election_group_id: SMALLINT -> INT
- user_practice_tests.election_group_id: SMALLINT -> INT (pgtype.Int4)
- user_referrals / referrals.election_group_id: SMALLINT -> INT (pgtype.Int4)
- agent_earnings.election_group_id: SMALLINT -> INT
- inec_result_grabber / inec_result_grabber_logs.election_group_id: SMALLINT -> INT

## sqlc-generated queries & models (apps/api/internal/db/queries/)
- models.go: update ElectionGroup, ElectionGroupID fields to int32 / pgtype.Int4
  across Election, PartyApplication, PollingUnitAssignment, PollingUnitResult,
  PollingUnitUpdate, Referral, UserReferral, UserPracticeTest,
  StateElectionSupervisor, LgaElectionSupervisor, WardElectionSupervisor,
  PartyElectionGroup, PartyMarketingCampaign, InecResultGrabberLog,
  UnmatchedPollingUnitResult
- election_groups.sql.go: GetElectionGroupByID / DeleteElectionGroup / UpdateElectionGroupParams.ID int16 -> int32
  UpsertPartyElectionGroupCoverageParams / UpsertPartyElectionGroupStatsParams.ElectionGroupID int16 -> int32
  ListElectionGroupsWithPartyStatsRow.ID int16 -> int32
- All other .sql.go files: param/row structs updated from int16/pgtype.Int2 to int32/pgtype.Int4

## Go handlers (apps/api/internal/handler/)
- election_groups: GetElectionGroupByID / DeleteElectionGroup URL param parsing int64->int32
- election_stats: all geo-level groupID parsing narrowed to int32
- elections / election_votes / non_voting: electionGroupID int32 propagation
- party_applications: SubmitApplicationRequest/SubmitSupervisorApplicationRequest.ElectionGroupID(s) int16->int32
  PartyApplicationsService interface updated (ListApplications, GetPollingUnitRecommendations) int16->int32
- polling_unit_assignments, polling_unit_results, polling_unit_updates, polling_units
- agent_earnings, agent_performance, practice_tests, referrals, supervisor_assignments
- parties, seed

## Go services (apps/api/internal/service/)
- election_groups, election_stats, elections, party_applications, polling_unit_assignments,
  polling_unit_results, polling_unit_updates, polling_units, earnings, referrals,
  supervisor_assignments: int16->int32 for all election_group_id fields and params

## Background workers (apps/api/internal/worker/)
- seed_election_stats_task.go, stats_geo_tasks.go: ElectionGroupID int32 propagation

## SQL query definitions (apps/api/db/query/)
- agent_performance, bodies, election_stats, final_results, party_applications,
  polling_unit_assignments, polling_unit_results, polling_unit_updates, practice_tests:
  cast and comparison types updated to ::integer / INT where relevant

+++ apps/api/db/migrations/20260607120000_elections_schema.sql
@@ -1,6 +1,6 @@
 -- +goose Up
 CREATE TABLE election_groups (
-  id SMALLINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+  id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   name VARCHAR(255) UNIQUE NOT NULL,
   rank INT NOT NULL,
   
@@ -90,7 +90,7 @@ CREATE TABLE election_groups (
 
 CREATE TABLE elections (
   id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-  election_group_id SMALLINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
+  election_group_id INT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
   state_id SMALLINT REFERENCES c_states(id) ON DELETE SET NULL,
   senatorial_district_id INT REFERENCES senatorial_districts(id) ON DELETE SET NULL,
   federal_constituency_id INT REFERENCES federal_constituencies(id) ON DELETE SET NULL,
