+++ apps/api/db/migrations/20260702200000_election_results.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS unmatched_polling_unit_results (
 
   -- Core context
   election_id              INT      REFERENCES elections(id) ON DELETE CASCADE NOT NULL,
-  election_group_id        SMALLINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
+  election_group_id INT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
   submitted_by             BIGINT   REFERENCES users(id) ON DELETE SET NULL,
   party_id                 SMALLINT REFERENCES parties(id) ON DELETE SET NULL,
 
@@ -68,7 +68,7 @@ CREATE TABLE polling_unit_results (
   -- Core relationships
   assignment_id            BIGINT   REFERENCES polling_unit_assignments(id) ON DELETE SET NULL,
   election_id              INT      REFERENCES elections(id) ON DELETE CASCADE NOT NULL,
-  election_group_id        SMALLINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
+  election_group_id INT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
   polling_unit_id          INTEGER  REFERENCES polling_units(id) ON DELETE CASCADE NOT NULL,
   submitted_by             BIGINT   REFERENCES users(id) ON DELETE SET NULL,
   party_id                 SMALLINT   REFERENCES parties(id) ON DELETE SET NULL,  -- NULL for general users
@@ -148,7 +148,7 @@ CREATE TABLE election_polling_unit_final_results (
   id                       BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
 
   election_id              INT      REFERENCES elections(id) ON DELETE CASCADE NOT NULL,
-  election_group_id        SMALLINT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
+  election_group_id INT REFERENCES election_groups(id) ON DELETE CASCADE NOT NULL,
   polling_unit_id          INTEGER  REFERENCES polling_units(id) ON DELETE CASCADE NOT NULL,
 
   -- Denormalized for fast geo-filtering
